requirements.txt
crewai==0.102.0
crewai_tools==0.36.0
pydantic==2.10.6
python-dotenv==1.0.1
streamlit==1.38.0
